requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfplumber
python-multipart
numpy
pydantic
